refactor: 도메인형 폴더 구조로 변경
