+++ src/main/java/doit/shop/domain/user/entity/User.java
@@ -1,0 +1,19 @@
+package doit.shop.domain.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String username;
+    private String password;
+    private String nickname;
+    private String phone;
+
+}
