+++ src/main/java/doit/shop/domain/user/service/UserService.java
@@ -1,0 +1,12 @@
+package doit.shop.domain.user.service;
+
+import doit.shop.domain.user.controller.UserController;
+import doit.shop.domain.user.dto.UserSignUpRequest;
+
+public class UserService {
+    private UserController userController;
+
+    public void signup(UserSignUpRequest userSignUpRequest) {
+
+    }
+}
