+++ src/main/java/doit/shop/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package doit.shop.common.config;
+package doit.shop.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
